Move serialized types to separate file
