+++ src/chord_sheet_serializer.ts
@@ -9,7 +9,16 @@ import Line from './chord_sheet/line';
 import AstType from './chord_sheet/ast_type';
 import Item from './chord_sheet/item';
 import Evaluatable from './chord_sheet/chord_pro/evaluatable';
-import { ChordType, Modifier } from './constants';
+
+import {
+  SerializedChordLyricsPair,
+  SerializedComment,
+  SerializedComponent,
+  SerializedItem,
+  SerializedLine, SerializedLiteral,
+  SerializedSong,
+  SerializedTag, SerializedTernary,
+} from './serialized_types';
 
 const CHORD_SHEET = 'chordSheet';
 const CHORD_LYRICS_PAIR = 'chordLyricsPair';
@@ -17,92 +26,6 @@ const TAG = 'tag';
 const COMMENT = 'comment';
 const TERNARY = 'ternary';
 const LINE = 'line';
-
-type SerializedTraceInfo = {
-  location?: {
-    offset: number | null,
-    line: number | null,
-    column: number | null,
-  },
-};
-
-type SerializedChord = {
-  type: 'chord',
-  base: string,
-  modifier: Modifier | null,
-  suffix: string | null,
-  bassBase: string | null,
-  bassModifier: Modifier | null,
-  chordType: ChordType,
-};
-
-export type SerializedChordLyricsPair = {
-  type: 'chordLyricsPair',
-  chord?: SerializedChord | null,
-  chords: string,
-  lyrics: string | null,
-  annotation?: string | null,
-};
-
-export type SerializedTag = SerializedTraceInfo & {
-  type: 'tag',
-  name: string,
-  value: string,
-};
-
-export type SerializedComment = {
-  type: 'comment',
-  comment: string,
-};
-
-export type ContentType = 'tab' | 'abc' | 'ly' | 'grid';
-
-export type SerializedSection = {
-  type: 'section',
-  sectionType: ContentType,
-  content: string[],
-  startTag: SerializedTag,
-  endTag: SerializedTag,
-};
-
-type SerializedLiteral = string;
-
-export interface SerializedTernary extends SerializedTraceInfo {
-  type: 'ternary',
-  variable: string | null,
-  valueTest: string | null,
-  trueExpression: Array<SerializedLiteral | SerializedTernary>,
-  falseExpression: Array<SerializedLiteral | SerializedTernary>,
-}
-
-export type SerializedComposite = Array<SerializedLiteral | SerializedTernary>;
-
-export type SerializedItem =
-  SerializedChordLyricsPair |
-  SerializedTag |
-  SerializedComment |
-  SerializedTernary |
-  SerializedLiteral;
-
-export type SerializedLine = {
-  type: 'line',
-  items: SerializedItem[],
-};
-
-export type SerializedSong = {
-  type: 'chordSheet',
-  lines: SerializedLine[],
-};
-
-type SerializedComponent =
-  SerializedLine |
-  SerializedSong |
-  SerializedChordLyricsPair |
-  SerializedTag |
-  SerializedComment |
-  SerializedTernary |
-  SerializedLiteral |
-  SerializedSection;
 
 /**
  * Serializes a song into een plain object, and deserializes the serialized object back into a {@link Song}
