+++ test/fixtures/serialized_song.ts
@@ -1,4 +1,4 @@
-import { SerializedSong } from '../../src/chord_sheet_serializer';
+import { SerializedSong } from '../../src/serialized_types';
 
 export const serializedSongSymbol: SerializedSong = {
   type: 'chordSheet',
