+++ test/formatter/html_table_formatter.test.ts
@@ -5,7 +5,8 @@ import '../matchers';
 import { exampleSongSolfege, exampleSongSymbol } from '../fixtures/song';
 import { scopedCss } from '../../src/formatter/html_table_formatter';
 import { stripHTML } from '../../src/template_helpers';
-import ChordSheetSerializer, { ContentType } from '../../src/chord_sheet_serializer';
+import ChordSheetSerializer from '../../src/chord_sheet_serializer';
+import { ContentType } from '../../src/serialized_types';
 
 import {
   chordLyricsPair, createSongFromAst, heredoc, html, section,
